add multiple option for select
when option "multiple" is true, will process original value as array
(joined on comma for now for display), and then pass array of selected
options directly back to controller.  Always adds a button because the
change event is often fired spuriously when interacting with a
multiselect.

fix fails because of asset precompilation and digests in 4.2

the block for Combustion.initialize! sets the old values for the test
env that turn off digests, so the matcher for the image tag in "should
be able to use bip_bool to change a boolean value using an image" works.

also moving css files to standard application.css manifest avoids the
complaints from rails about "Asset was not declared to be precompiled in
production." for each of these.

defend against nil values for given field

travis kick

smack travis

travis kick harder